+++ lib/cucumber/grape.rb
@@ -1,5 +1,5 @@
 module Cucumber
-  module Sinatra
+  module Grape
     VERSION = '0.5.0'
   end
 end